* scrap the auto-redirect and just show that they need to connect their wallet
