+++ lib/components/AccountSummary.jsx
@@ -1,13 +1,12 @@
 import React from 'react'
 import classnames from 'classnames'
-import FeatherIcon from 'feather-icons-react'
 import { useRouter } from 'next/router'
 import { useAtom } from 'jotai'
 import { useOnboard } from '@pooltogether/hooks'
-import { Button, Tooltip } from '@pooltogether/react-components'
 import { useTranslation } from 'react-i18next'
 
 import { isSelfAtom } from 'lib/components/AccountUI'
+import { GenericConnectWalletCTA } from 'lib/components/GenericConnectWalletCTA'
 import { PoolNumber } from 'lib/components/PoolNumber'
 import { ThemedClipSpinner } from 'lib/components/loaders/ThemedClipSpinner'
 import { usePlayerTotalDepositValue } from 'lib/hooks/useUserTickets'
@@ -21,7 +20,7 @@ export const AccountSummary = () => {
 
   const [isSelf] = useAtom(isSelfAtom)
 
-  const { connectWallet, address: usersAddress } = useOnboard()
+  const { address: usersAddress } = useOnboard()
 
   const router = useRouter()
   const playerAddress = router?.query?.playerAddress
@@ -49,58 +48,7 @@ export const AccountSummary = () => {
               $<PoolNumber>{numberWithCommas(totalValueUsd, { precision: 2 })}</PoolNumber>
             </h1>
           ) : !address ? (
-            <>
-              <div className='text-xxxs sm:text-sm font-bold mt-1'>
-                {t('connectAWalletToManageTicketsAndRewards')}
-              </div>
-
-              <Button
-                textSize='xs'
-                padding='px-4 py-1'
-                className='mt-4'
-                onClick={(e) => {
-                  e.preventDefault()
-                  connectWallet(() => {})
-                }}
-              >
-                {t('connectWallet')}
-              </Button>
-
-              <div className='flex flex-col sm:flex-row sm:items-center mt-8'>
-                <Tooltip
-                  id='what-is-ethereum-tooltip'
-                  tip={
-                    <>
-                      Ethereum Wallet?
-                      <br />
-                      {t('whatIsEthereumOne')}
-                      <br />
-                      <br />
-                      {t('whatIsEthereumTwo')}
-                    </>
-                  }
-                >
-                  <div className='flex items-center opacity-60 font-bold text-accent-1 text-xxs'>
-                    <FeatherIcon icon='info' className='w-3 h-3 sm:w-4 sm:h-4 inline-block mr-1' />
-                    {t('whatsAnEthereum')}
-                  </div>
-                </Tooltip>
-
-                <div className='opacity-60 font-bold sm:ml-8 mt-1 sm:mt-0'>
-                  <a
-                    href='https://ethereum.org/en/wallets'
-                    target='_blank'
-                    className='text-accent-1 text-xxxxs sm:text-xxs flex items-center'
-                  >
-                    {t('learnMoreAboutEthereumWallets')}{' '}
-                    <FeatherIcon
-                      icon='arrow-up-right'
-                      className='w-3 h-3 sm:w-4 sm:h-4 inline-block ml-1'
-                    />
-                  </a>
-                </div>
-              </div>
-            </>
+            <GenericConnectWalletCTA />
           ) : (
             <ThemedClipSpinner />
           )}
