+++ pages/account/pools/[networkName]/[symbol]/manage-tickets.jsx
@@ -1,15 +1,12 @@
 import React from 'react'
 
-import { AccountLoggedIn } from 'lib/components/AccountLoggedIn'
 import { AccountUI } from 'lib/components/AccountUI'
 import Layout from 'lib/components/Layout'
 
 export default function AccountManage(props) {
   return (
     <Layout>
-      <AccountLoggedIn>
-        <AccountUI />
-      </AccountLoggedIn>
+      <AccountUI />
     </Layout>
   )
 }
